move: Auction DTO Response 폴더 생성으로 인한 파일 이동
(#61)
