+++ src/main/java/org/chzz/market/domain/auction/repository/AuctionRepositoryImpl.java
@@ -22,11 +22,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.chzz.market.domain.auction.dto.AuctionDetailsResponse;
-import org.chzz.market.domain.auction.dto.AuctionResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionDetailsResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionResponse;
 import org.chzz.market.domain.auction.dto.QAuctionDetailsResponse;
 import org.chzz.market.domain.auction.dto.QAuctionResponse;
-import org.chzz.market.domain.auction.entity.Auction.Status;
 import org.chzz.market.domain.auction.entity.SortType;
 import org.chzz.market.domain.image.entity.QImage;
 import org.chzz.market.domain.product.entity.Product.Category;
