+++ src/test/java/org/chzz/market/domain/auction/service/AuctionServiceTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.when;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
-import org.chzz.market.domain.auction.dto.AuctionDetailsResponse;
+import org.chzz.market.domain.auction.dto.response.AuctionDetailsResponse;
 import org.chzz.market.domain.auction.entity.Auction;
 import org.chzz.market.domain.auction.error.AuctionException;
 import org.chzz.market.domain.auction.repository.AuctionRepository;
